[Refactor] refactor copy and create link cards
